Check `allow_recursive` on the watcher list.
The default value is true, which gives the same behavior as before.
When the value is false, the ``addresses`` property only looks for
watchers on the current item, not on the parent.

A sample usage would be in Products.Poi to only allow recursive if an
issue is not yet assigned.  (I am going to use this in custom code for
a client).